requirements.txt
python-telegram-bot==12.7
requests==2.28.1
APScheduler==3.9.0